moved components into their own subfolder
