+++ todo-list/src/components/Current.js
@@ -2,7 +2,7 @@ import TodoList from "./TodoList";
 import { useNavigate } from "react-router-dom";
 import AddTodoForm from "./AddTodoForm";
 import NavButton from "./NavButton";
-import styles from "./Assets/css/App.module.css"
+import styles from "../Assets/css/App.module.css"
 
 const CurrentList = ({ todoList, addTodo, currentUser, onDone }) => {
 	let navigate = useNavigate();
