+++ todo-list/src/components/NavButton.js
@@ -1,4 +1,4 @@
-import styles from "./Assets/css/App.module.css"
+import styles from "../Assets/css/App.module.css"
 
 const NavButton = ({ type, action, children, path }) => {
 	const handleClick = (path) => {
